requirements.txt
aiomysql
pymysql
fastapi[all]
scalar_fastapi
pydantic-settings
sqlmodel
alembic
dotenv
passlib[bcrypt]
bcrypt==4.1.2
pyjwt